+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.1]
+
+### Fixed
+
+- Fixed broken unit test.
+
 ## [0.1.0]
 
 ### Fixed
@@ -29,7 +35,7 @@ def walk(root):
 
     Yields:
         Active nodes.
-    
+
     """
     def skip(node):
         return node.inactive